Refactor ext4 code of setup-mender-configuration into a separate file
So that the functions can be imported from other scripts. This commit
does not add actual changes in the code.

Changelog: None
